GH-877: Consolidate logic for protoc plugin resolution into one class
- Merge JvmPluginResolver and BinaryPluginResolver into ProjectPluginResolver,
  and rename to ProtocPluginResolver.
- Implicitly resolve plugins in parallel, fixing a long-standing TODO.
